Refactor codebase to group similar functions and limit file size
