+++ app/templates.py
@@ -1,0 +1,3 @@
+from fastapi.templating import Jinja2Templates
+
+templates = Jinja2Templates(directory="templates")
